% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R, R/xrays.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{mutate}
\alias{transmute}
\alias{select}
\alias{rename}
\alias{pull}
\alias{slice}
\alias{group_by}
\alias{summarise}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:mutate]{transmute}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

